refactor: type 병합
